% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{cl_attachments}
\alias{cl_attachments}
\title{Expand chat attachments into a table}
\usage{
cl_attachments(chat)
}
\arguments{
\item{chat}{A `chatlens_chat` object}
}
\description{
Expand chat attachments into a table
}
