% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{cl_chat_filter_period}
\alias{cl_chat_filter_period}
\title{Filter chat by period key}
\usage{
cl_chat_filter_period(chat, period, key)
}
\arguments{
\item{chat}{A `chatlens_chat` object}

\item{period}{One of `"year"`, `"month"`, `"week"`, or `"day"`}

\item{key}{Period key value to keep}
}
\description{
Filter chat by period key
}
