% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{cl_chat_split_periods}
\alias{cl_chat_split_periods}
\title{Slice chat text by time period}
\usage{
cl_chat_split_periods(
  chat,
  period = c("all", "year", "month", "week", "day"),
  text_col = "text_enriched"
)
}
\arguments{
\item{chat}{A `chatlens_chat`}

\item{period}{One or more of `"all"`, `"year"`, `"month"`, `"week"`, `"day"`}

\item{text_col}{Which text column to use}
}
\description{
Slice chat text by time period
}
