% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{cl_detect_participants}
\alias{cl_detect_participants}
\title{Detect participants in a chat}
\usage{
cl_detect_participants(chat)
}
\arguments{
\item{chat}{A `chatlens_chat` object}
}
\description{
Detect participants in a chat
}
