% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{cl_filter_date}
\alias{cl_filter_date}
\title{Filter chat by date range}
\usage{
cl_filter_date(chat, start = NULL, end = NULL)
}
\arguments{
\item{chat}{A `chatlens_chat` object}

\item{start}{Optional inclusive start datetime}

\item{end}{Optional inclusive end datetime}
}
\description{
Filter chat by date range
}
