% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{cl_whatsapp_summary}
\alias{cl_whatsapp_summary}
\title{Summarize an imported WhatsApp chat}
\usage{
cl_whatsapp_summary(chat = NULL, path = NULL, ask_confirmation = FALSE, ...)
}
\arguments{
\item{chat}{A \code{chatlens_chat} object}

\item{path}{Optional ZIP path to import and then summarize}

\item{ask_confirmation}{Ask to confirm detected chat file}

\item{...}{Additional arguments passed to \code{\link[=cl_whatsapp_import]{cl_whatsapp_import()}}}
}
\description{
Summarize an imported WhatsApp chat
}
