% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{cl_chat_anonymize}
\alias{cl_chat_anonymize}
\title{Anonymize a chat using an alias map}
\usage{
cl_chat_anonymize(
  chat,
  alias_map = NULL,
  interactive = FALSE,
  save_original = TRUE,
  cache_dir = cl_cache_dir()
)
}
\arguments{
\item{chat}{A chatlens_chat}

\item{alias_map}{Alias map (optional)}

\item{interactive}{Prompt for aliases}

\item{save_original}{Save a copy of the original chat in cache}

\item{cache_dir}{Cache directory}
}
\description{
Anonymize a chat using an alias map
}
