% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{cl_chat_process_media}
\alias{cl_chat_process_media}
\title{Enrich text with media annotations}
\usage{
cl_chat_process_media(
  chat,
  include_audio = TRUE,
  include_images = TRUE,
  audio_tag = "AUDIO TRANSCRIPT",
  image_tag = "IMAGE DESCRIPTION"
)
}
\arguments{
\item{chat}{A `chatlens_chat` object}

\item{include_audio}{Include audio transcript annotations}

\item{include_images}{Include image description annotations}

\item{audio_tag}{Tag label used for audio annotations}

\item{image_tag}{Tag label used for image annotations}
}
\description{
Enrich text with media annotations
}
